adds rules support to be applied to a node; improves node properites

+++ pkg/model/model.go
@@ -73,7 +73,7 @@ type Disk struct {
 }
 
 type Memory struct {
-	PhysicalMb float32 `json:"physical_mb"`
+	PhysicalMb int     `json:"physical_mb"`
 	Total      float32 `json:"total"`
 }
